+++ pyan/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-__version__ = "2.4.4-dev"
+__version__ = "2.5.0-dev"
 
 from .main import create_callgraph, main  # noqa: F401
 from .modvis import create_modulegraph  # noqa: F401